merge

+++ src/main/java/site/alice/liveman/web/rpc/BroadcastController.java
@@ -143,18 +143,14 @@ public class BroadcastController {
         }
         VideoInfo videoInfo = mediaProxyTask.getVideoInfo();
         BroadcastTask broadcastTask = videoInfo.getBroadcastTask();
-        if (broadcastTask == null) {
-            log.info("此转播任务尚未运行，或已停止[BroadcastTask不存在][videoId=" + videoId + "]");
-            return ActionResult.getErrorResult("此转播任务尚未运行或已停止");
-        }
-        AccountInfo broadcastAccount = broadcastTask.getBroadcastAccount();
-        if (broadcastAccount == null) {
-            log.info("此转播任务尚未运行，或已停止[BroadcastAccount不存在][videoId=" + videoId + "]");
-            return ActionResult.getErrorResult("此转播任务尚未运行或已停止");
-        }
-        if (!broadcastAccount.getRoomId().equals(account.getRoomId()) && !account.isAdmin()) {
-            log.info("您没有权限修改他人直播间的推流任务[videoId=" + videoId + "][broadcastRoomId=" + broadcastAccount.getRoomId() + "]");
-            return ActionResult.getErrorResult("你没有权限修改他人直播间的推流任务");
+        if (broadcastTask != null) {
+            AccountInfo broadcastAccount = broadcastTask.getBroadcastAccount();
+            if (broadcastAccount != null) {
+                if (!broadcastAccount.getRoomId().equals(account.getRoomId()) && !account.isAdmin()) {
+                    log.info("您没有权限编辑他人直播间的推流任务[videoId=" + videoId + "][broadcastRoomId=" + broadcastAccount.getRoomId() + "]");
+                    return ActionResult.getErrorResult("你没有权限编辑他人直播间的推流任务");
+                }
+            }
         }
         if (cropConf != null) {
             if (cropConf.getVideoBannedType() == VideoBannedTypeEnum.AREA_SCREEN && !account.isVip()) {
@@ -168,7 +164,9 @@ public class BroadcastController {
                 log.error("保存系统配置信息失败", e);
                 return ActionResult.getErrorResult("系统内部错误，请联系管理员");
             }
-            ProcessUtil.killProcess(broadcastTask.getPid());
+            if (broadcastTask != null) {
+                ProcessUtil.killProcess(broadcastTask.getPid());
+            }
         }
         return ActionResult.getSuccessResult(null);
     }
@@ -248,19 +246,17 @@ public class BroadcastController {
         BroadcastTask broadcastTask = videoInfo.getBroadcastTask();
         if (broadcastTask != null) {
             AccountInfo broadcastAccount = broadcastTask.getBroadcastAccount();
-            if (broadcastAccount == null) {
-                log.info("此转播任务尚未运行，或已停止[BroadcastAccount不存在][videoId=" + videoId + "]");
-                return ActionResult.getErrorResult("此转播任务尚未运行或已停止");
+            if (broadcastAccount != null) {
+                if (!broadcastAccount.getRoomId().equals(account.getRoomId()) && !account.isAdmin()) {
+                    log.info("您没有权限编辑他人直播间的推流任务[videoId=" + videoId + "][broadcastRoomId=" + broadcastAccount.getRoomId() + "]");
+                    return ActionResult.getErrorResult("你没有权限编辑他人直播间的推流任务");
+                }
+                Integer areaId = null;
+                if (broadcastTaskVO.getArea() != null && broadcastTaskVO.getArea().length > 1) {
+                    areaId = broadcastTaskVO.getArea()[1];
+                }
+                broadcastServiceManager.getBroadcastService(account.getAccountSite()).setBroadcastSetting(broadcastAccount, broadcastTaskVO.getRoomTitle(), areaId);
             }
-            if (!broadcastAccount.getRoomId().equals(account.getRoomId()) && !account.isAdmin()) {
-                log.info("您没有权限编辑他人直播间的推流任务[videoId=" + videoId + "][broadcastRoomId=" + broadcastAccount.getRoomId() + "]");
-                return ActionResult.getErrorResult("你没有权限编辑他人直播间的推流任务");
-            }
-            Integer areaId = null;
-            if (broadcastTaskVO.getArea() != null && broadcastTaskVO.getArea().length > 1) {
-                areaId = broadcastTaskVO.getArea()[1];
-            }
-            broadcastServiceManager.getBroadcastService(account.getAccountSite()).setBroadcastSetting(broadcastAccount, broadcastTaskVO.getRoomTitle(), areaId);
         }
         videoInfo.setArea(broadcastTaskVO.getArea());
         videoInfo.setNeedRecord(broadcastTaskVO.isNeedRecord());
